backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Supabase
supabase==2.3.0

# Authentication & Security
pyjwt==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# HTTP Client (version compatible with Supabase 2.3.0)
httpx>=0.24.0,<0.25.0

# Image Processing
pillow==10.1.0
pytesseract==0.3.10
google-cloud-vision==3.7.0
rapidfuzz==3.6.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# LLM Integration
openai>=1.3.0
google-generativeai>=0.3.0

# Environment Management
python-dotenv==1.0.0
